[CI Visibility] Align CODEOWNERS glob semantics and honor useOSSeparator when anchoring
- Honor useOSSeparator in TryAnchorPathToCodeOwnersRoot
- Treat ? as a single non-slash character and ** as a globstar only when it is a whole path segment
- Drop unused lineNo and fix the CompileGlob comment
- Move fixture-based CodeOwnersTests into the unit-test project
